+++ example/lib/presentation/src/ac_calendar_scope.dart
@@ -10,21 +10,28 @@ import 'theme/theme.dart';
 /// Оборачивает дочерний виджет в [ACCalendarTheme], обеспечивая совместимость
 /// с обоими механизмами получения темы — через scope и напрямую.
 class ACCalendarScope extends InheritedWidget {
-  ACCalendarScope({
-    required this.theme,
-    required this.dateRange,
+  factory ACCalendarScope({
+    required ACDateRange dateRange,
     required Widget child,
-    this.selectController,
-    super.key,
-  }) : super(
+    ACCalendarThemeData? theme,
+    ACCalendarSelectController? selectController,
+    Key? key,
+  }) => ACCalendarScope.raw(
+    dateRange: dateRange,
+    selectController: selectController,
+    key: key,
     child: ACCalendarTheme(
-      data: theme,
-      child: child
-    )
+      data: theme ?? ACLightCalendarThemeData(),
+      child: child,
+    ),
   );
 
-  /// Тема календаря.
-  final ACCalendarThemeData theme;
+  const ACCalendarScope.raw({
+    required this.dateRange,
+    required super.child,
+    this.selectController,
+    super.key,
+  });
 
   /// Диапазон допустимых дат календаря.
   final ACDateRange dateRange;
@@ -43,7 +50,6 @@ class ACCalendarScope extends InheritedWidget {
 
   @override
   bool updateShouldNotify(ACCalendarScope oldWidget) =>
-    theme != oldWidget.theme ||
     dateRange != oldWidget.dateRange ||
     selectController != oldWidget.selectController;
 }