+++ example/lib/presentation/src/theme/theme.dart
@@ -1,6 +1,8 @@
-export 'src/ac_calendar_header_theme_data.dart';
 export 'src/ac_calendar_theme.dart';
 export 'src/ac_calendar_theme_data.dart';
 export 'src/ac_day_theme_data.dart';
+export 'src/ac_month_picker_theme_data.dart';
+export 'src/ac_pages_calendar_header_theme_data.dart';
+export 'src/ac_titled_month_theme_data.dart';
 export 'src/ac_week_theme_data.dart';
-export 'src/ac_wheel_theme_data.dart';
+export 'src/ac_wheel_picker_theme_data.dart';