+++ example/lib/presentation/src/widgets/src/calendar/src/vertical/vertical.dart
@@ -1,2 +1,0 @@
-export 'src/ac_calendar_layout.dart';
-export 'src/ac_calendar_widget.dart';