+++ example/lib/presentation/src/widgets/src/scroll_view/scroll_view.dart
@@ -1,2 +1,3 @@
+export 'src/ac_date_range_scroll_view_controller.dart';
 export 'src/ac_scroll_view.dart';
 export 'src/ac_scroll_view_controller.dart';